added some editor commands

+++ src/filesystem.h
@@ -4,6 +4,6 @@
 #include "buffer.h"
 
 bool load_doc_data(struct Document *doc, char *path);
-int save_doc(struct Document *doc);
+size_t save_doc(struct Document *doc);
 
 #endif
